+++ packages/agents-a365-observability/src/tracing/scopes/InferenceScope.ts
@@ -96,7 +96,7 @@ export class InferenceScope extends OpenTelemetryScope {
    * @param messages Array of input messages
    */
   public recordInputMessages(messages: string[]): void {
-    this.setTagMaybe(OpenTelemetryConstants.GEN_AI_INPUT_MESSAGES_KEY, messages.join(','));
+    this.setTagMaybe(OpenTelemetryConstants.GEN_AI_INPUT_MESSAGES_KEY, JSON.stringify(messages));
   }
 
   /**
@@ -104,7 +104,7 @@ export class InferenceScope extends OpenTelemetryScope {
    * @param messages Array of output messages
    */
   public recordOutputMessages(messages: string[]): void {
-    this.setTagMaybe(OpenTelemetryConstants.GEN_AI_OUTPUT_MESSAGES_KEY, messages.join(','));
+    this.setTagMaybe(OpenTelemetryConstants.GEN_AI_OUTPUT_MESSAGES_KEY, JSON.stringify(messages));
   }
 
   /**